# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  full_name?: string;
  password?: string;
}